requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-amazon==7.1.0
pytest
pytest-cov
